[BE] refactor : 폴더구조 도메인형으로  변경, reservation request 타입 변경에 따른 reservation service 내에서의 reservation builder 리팩토링
