+++ backend/driving-today/src/main/java/com/drivingtoday/domain/academy/Academy.java
@@ -1,4 +1,4 @@
-package com.drivingtoday.entity;
+package com.drivingtoday.domain.academy;
 
 
 import jakarta.persistence.*;
