+++ backend/driving-today/src/main/java/com/drivingtoday/domain/reservation/ReservationController.java
@@ -1,12 +1,8 @@
-package com.drivingtoday.controller;
+package com.drivingtoday.domain.reservation;
 
 
-import com.drivingtoday.DTO.ReservationDTO;
-import com.drivingtoday.entity.Reservation;
-import com.drivingtoday.service.ReservationService;
+import com.drivingtoday.domain.reservation.dto.ReservationRequest;
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.Parameters;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -23,7 +19,7 @@ public class ReservationController {
 
     @PostMapping("/")
     @Operation(summary = "예약만들기 API")
-    public ResponseEntity<?> createReservation(@RequestBody ReservationDTO reservationRequest){
+    public ResponseEntity<?> createReservation(@RequestBody ReservationRequest reservationRequest){
         reservationService.saveReservations(reservationRequest);
 
         return ResponseEntity.ok().build();
