+++ backend/driving-today/src/main/java/com/drivingtoday/domain/review/Review.java
@@ -1,6 +1,8 @@
-package com.drivingtoday.entity;
+package com.drivingtoday.domain.review;
 
 
+import com.drivingtoday.domain.instructor.Instructor;
+import com.drivingtoday.domain.student.Student;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
