+++ backend/driving-today/src/main/java/com/drivingtoday/domain/review/ReviewRepository.java
@@ -1,6 +1,6 @@
-package com.drivingtoday.repository;
+package com.drivingtoday.domain.review;
 
-import com.drivingtoday.entity.Review;
+import com.drivingtoday.domain.review.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReviewRepository extends JpaRepository<Review, Long> {
