--- backend/driving-today/src/main/java/com/drivingtoday/entity/Academy.java
@@ -1,33 +1,0 @@
-package com.drivingtoday.entity;
-
-
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotNull;
-import lombok.*;
-
-@Entity
-@NoArgsConstructor
-@Getter
-@Builder
-@AllArgsConstructor
-public class Academy {
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "academy_id")
-    private Long id;
-
-    @Column(name = "name")
-    @NotNull
-    private String name;
-
-    @Column(name = "latitude")
-    @NotNull
-    private Double latitude;
-
-    @Column(name = "longitude")
-    @NotNull
-    private Double longitude;
-
-    @Column(name = "cert")
-    @NotNull
-    private Boolean cert;
-}
