+++ backend/driving-today/src/main/java/com/drivingtoday/global/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.drivingtoday.config;
+package com.drivingtoday.global.config;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
